New pre-release: 1.1-alpha1
Added markdown support, debug and stylish. :)
Feature: Markdown Support
Debug: (Backend) Fixed Compilers Manager (Java Support).
Debug: (Backend) Added Problem 1001 to test the scoring system.
