--- client_py/P1000.py
@@ -1,1 +1,0 @@
-print("Genshin Impact, launch!")
