+++ compilers_manager/compilers/java_compiler.py
@@ -1,4 +1,4 @@
-import os
+import os, json, asyncio
 
 import compilers_manager.compilers.base_compiler
 
@@ -7,9 +7,26 @@ class LanguageNotSupportedException(Exception):
     pass
 
 
+async def execute_command(command: str, timeout: int | float | None = None):
+    proc = await asyncio.create_subprocess_shell(
+        command, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE
+    )
+    try:
+        async with asyncio.timeout(timeout):
+            stdout, stderr = await proc.communicate()
+            print(f"{command!r} exited with {proc.returncode}, pid={proc.pid}")
+            return proc.returncode
+    except TimeoutError:
+        proc.terminate()
+        raise TimeoutError
+    except Exception as e:
+        proc.terminate()
+        raise e
+
+
 class java_compiler(compilers_manager.compilers.base_compiler.base_compiler):
 
-    def __init__(self, unload_timeout) -> None:
+    def __init__(self, unload_timeout: int) -> None:
         print("Java Compiler loaded.")
         __slots__ = (
             "__init__",
@@ -26,41 +43,62 @@ class java_compiler(compilers_manager.compilers.base_compiler.base_compiler):
         print("Java Compiler unloaded.")
 
     async def on_compile(
-        self, language, compile_file_path, compile_binary_path
+        self, language: str, compile_file_path: str, compile_binary_path: str
     ) -> bool:
         if language == "java":
             try:
-                os.system("javac {}".format(compile_file_path, compile_binary_path))
-            except OSError:
-                pass
+                compile_file_name_without_extension, _ = os.path.splitext(
+                    compile_file_path
+                )
+                compile_file_name_without_extension = (
+                    compile_file_name_without_extension.split("/")[-1]
+                )
+                print(compile_file_path, compile_file_name_without_extension)
+                with open(compile_file_path, "a+") as compile_file:
+                    compile_file.writelines(
+                        [
+                            "\n",
+                            "public class "
+                            + compile_file_name_without_extension
+                            + "{\n",
+                            "    public static void main(String[] args) {\n",
+                            "        new MyJavaSourceCode();\n",
+                            "    }\n",
+                            "}\n",
+                        ]
+                    )
+                    compile_file.flush()
+
+                exit_code = await execute_command("javac {}".format(compile_file_path))
+                return exit_code == 0 and os.path.exists(compile_binary_path)
             except:
                 return False
-
-            if not os.path.exists(compile_binary_path):
-                return False
-
-            return True
         else:
             raise LanguageNotSupportedException(
                 "The language {} is not supported.".format(language)
             )
 
     async def on_cleanup(
-        self, language, compile_file_path, compile_binary_path
+        self, language: str, compile_file_path: str, compile_binary_path: str
     ) -> bool:
         if language == "java":
             try:
                 os.remove(compile_file_path)
                 os.remove(compile_binary_path)
-            except OSError:
+                os.remove(
+                    os.path.dirname(compile_file_path) + "/MyJavaSourceCode.class"
+                )
                 pass
             except:
                 return False
 
-            if os.path.exists(compile_file_path) or os.path.exists(compile_binary_path):
-                return False
-
-            return True
+            return not (
+                os.path.exists(compile_file_path)
+                or os.path.exists(compile_binary_path)
+                or os.path.exists(
+                    os.path.dirname(compile_file_path) + "/MyJavaSourceCode.class"
+                )
+            )
         else:
             raise LanguageNotSupportedException(
                 "The language {} is not supported.".format(language)
@@ -76,7 +114,7 @@ class java_compiler(compilers_manager.compilers.base_compiler.base_compiler):
 
     def get_binary_extension(self, language: str) -> str:
         if language == "java":
-            return ".javaw"
+            return ".class"
         else:
             raise LanguageNotSupportedException(
                 "The language {} is not supported.".format(language)
@@ -86,7 +124,14 @@ class java_compiler(compilers_manager.compilers.base_compiler.base_compiler):
         self, language: str, compile_file_path: str
     ) -> str:
         if language == "java":
-            return "java {}.javaw".format(compile_file_path)
+            compile_file_name_without_extension, _ = os.path.splitext(compile_file_path)
+            compile_file_name_without_extension = (
+                compile_file_name_without_extension.split("/")[-1]
+            )
+            return "java -classpath {} {}".format(
+                os.path.dirname(compile_file_path),
+                compile_file_name_without_extension,
+            )
         else:
             raise LanguageNotSupportedException(
                 "The language {} is not supported.".format(language)
