+++ problem/1000/problem_statement.json
@@ -3,6 +3,7 @@
     "difficulty": 1,
     "problem_name": "Genshin Impact",
     "problem_statement": [
-        "You' re given one integer and you should print \"Genshin Impact, launch!\"."
+        "You' re given one integer and you should print \"Genshin Impact, launch!\".",
+        "**Note that at the end of the output, you should print a line feed, or your answer may be judged to be wrong.** :("
     ]
 }
