+++ problem/1001/answer/answer1.txt
@@ -1,0 +1,1 @@
+3
