+++ problem/1001/answer/answer10.txt
@@ -1,0 +1,1 @@
+6912
