+++ problem/1001/answer/answer3.txt
@@ -1,0 +1,1 @@
+9090
