+++ problem/1001/answer/answer5.txt
@@ -1,0 +1,1 @@
+946
