+++ problem/1001/answer/answer6.txt
@@ -1,0 +1,1 @@
+88888
