+++ problem/1001/answer/answer7.txt
@@ -1,0 +1,1 @@
+1008244354
