+++ problem/1001/answer/answer8.txt
@@ -1,0 +1,1 @@
+17688
