+++ problem/1001/answer/answer9.txt
@@ -1,0 +1,1 @@
+2400
