+++ problem/1001/input/input1.txt
@@ -1,0 +1,1 @@
+1 2
