+++ problem/1001/input/input10.txt
@@ -1,0 +1,1 @@
+1234 5678
