+++ problem/1001/input/input3.txt
@@ -1,0 +1,1 @@
+4545 4545
