+++ problem/1001/input/input4.txt
@@ -1,0 +1,1 @@
+1111 1222
