+++ problem/1001/input/input6.txt
@@ -1,0 +1,1 @@
+44444 44444
