+++ problem/1001/input/input7.txt
@@ -1,0 +1,1 @@
+10000001 998244353
