+++ problem/1001/input/input8.txt
@@ -1,0 +1,1 @@
+7788 9900
