+++ problem/1001/input/input9.txt
@@ -1,0 +1,1 @@
+1200 1200
