--- submit/submission_1.cpp
@@ -1,1 +1,0 @@
-#include <bits/stdc++.h>using namespace std;signed main(){    cout << "Genshin Impact, launch!" << endl;    return 0;}
