+++ test_code/P1000.c
@@ -1,0 +1,7 @@
+#include <stdio.h>
+
+signed main()
+{
+	printf("Genshin Impact, launch!\n");
+	return 0;
+}
