+++ test_code/P1000.java
@@ -1,0 +1,5 @@
+class MyJavaSourceCode {
+    public MyJavaSourceCode() {
+        System.out.println("Genshin Impact, launch!"); 
+    }
+}
