+++ test_code/P1000.py
@@ -1,0 +1,1 @@
+print("Genshin Impact, launch!")
