+++ test_code/P1001.c
@@ -1,0 +1,9 @@
+#include <stdio.h>
+
+signed main()
+{
+    long long A, B;
+    scanf("%lld %lld", &A, &B);
+    printf("%lld\n", A + B);
+    return 0;
+}
