+++ test_code/P1001.cpp
@@ -1,0 +1,10 @@
+#include <iostream>
+using namespace std;
+
+signed main()
+{
+    long long A, B;
+    cin >> A >> B;
+    cout << A + B << endl;
+    return 0;
+}
