+++ test_code/P1001.java
@@ -1,0 +1,10 @@
+import java.util.Scanner;
+
+class MyJavaSourceCode {
+    public MyJavaSourceCode() {
+        Scanner in = new Scanner(System.in);
+        int a = in.nextInt();
+        int b = in.nextInt();
+        System.out.println(a + b); 
+    }
+}
