+++ test_code/P1001.py
@@ -1,0 +1,4 @@
+a, b = input().split()
+a = int(a)
+b = int(b)
+print(a + b)
